requirements.txt
# Flask Web Framework
Flask==3.0.0
Werkzeug==3.0.1

# Google Cloud Services  
google-cloud-storage==2.10.0
google-cloud-firestore==2.13.1

# Google AI
google-genai==0.8.0

# Data Processing
pandas==2.1.4
numpy==1.24.3

# Image Processing
Pillow==10.1.0

# Environment Management
python-dotenv==1.0.0

# HTTP Requests
requests==2.31.0

# Additional utilities
uuid==1.30

# Production WSGI Server
gunicorn==21.2.0

# Optional: For enhanced logging in production
google-cloud-logging==3.8.0 